Extract the `status` router field explicitly
When Splunk automatically extracts fields for a request such as:

```
path="/wp-admin/admin.php?page=download_report&report=users&status=all"
```

It will extract the `path` field, and also extract `page`, `report` and `status` as fields out of the value. For `status` this is unexpected as there's another field later on in the message that's more important for us to extract.

Being explicit about extracting the `status` field resolves the problem.